# execution: refine precision of cast as decimal in agg func