--- a/cmd/explaintest/r/explain_easy.result
+++ b/cmd/explaintest/r/explain_easy.result
@@ -846,3 +846,118 @@ Projection	8000.00	root		Column#4, Column#4, Column#5
           └─HashAgg	8000.00	cop[tikv]		group by:test.t.b, funcs:sum(test.t.a)->Column#13, funcs:count(test.t.a)->Column#14
             └─TableFullScan	10000.00	cop[tikv]	table:t	keep order:false, stats:pseudo
 drop table if exists t;
+create table t(a tinyint, b smallint, c mediumint, d int, e bigint);
+insert into mysql.opt_rule_blacklist VALUES("aggregation_push_down");
+admin reload opt_rule_blacklist;
+
+explain format = 'brief' select sum(t1.a) from t t1 join t t2 on t1.a=t2.a;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.a, decimal(25,0) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.a, test.t.a)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.a))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.a))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+explain format = 'brief' select sum(t1.b) from t t1 join t t2 on t1.b=t2.b;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.b, decimal(27,0) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.b, test.t.b)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.b))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.b))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+explain format = 'brief' select sum(t1.c) from t t1 join t t2 on t1.c=t2.c;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.c, decimal(30,0) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.c, test.t.c)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.c))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.c))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+explain format = 'brief' select sum(t1.d) from t t1 join t t2 on t1.d=t2.d;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.d, decimal(32,0) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.d, test.t.d)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.d))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.d))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+explain format = 'brief' select sum(t1.e) from t t1 join t t2 on t1.e=t2.e;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.e, decimal(41,0) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.e, test.t.e)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.e))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.e))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+explain format = 'brief' select avg(t1.a) from t t1 join t t2 on t1.a=t2.a;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:avg(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.a, decimal(8,4) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.a, test.t.a)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.a))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.a))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+explain format = 'brief' select avg(t1.b) from t t1 join t t2 on t1.b=t2.b;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:avg(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.b, decimal(10,4) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.b, test.t.b)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.b))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.b))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+explain format = 'brief' select avg(t1.c) from t t1 join t t2 on t1.c=t2.c;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:avg(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.c, decimal(13,4) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.c, test.t.c)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.c))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.c))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+explain format = 'brief' select avg(t1.d) from t t1 join t t2 on t1.d=t2.d;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:avg(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.d, decimal(15,4) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.d, test.t.d)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.d))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.d))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+explain format = 'brief' select avg(t1.e) from t t1 join t t2 on t1.e=t2.e;
+id	estRows	task	access object	operator info
+StreamAgg	1.00	root		funcs:avg(Column#14)->Column#13
+└─Projection	12487.50	root		cast(test.t.e, decimal(24,4) BINARY)->Column#14
+  └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.e, test.t.e)]
+    ├─TableReader(Build)	9990.00	root		data:Selection
+    │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.e))
+    │   └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
+    └─TableReader(Probe)	9990.00	root		data:Selection
+      └─Selection	9990.00	cop[tikv]		not(isnull(test.t.e))
+        └─TableFullScan	10000.00	cop[tikv]	table:t1	keep order:false, stats:pseudo
+drop table if exists t;