--- a/cmd/explaintest/r/explain_easy.result
+++ b/cmd/explaintest/r/explain_easy.result
@@ -91,7 +91,7 @@ Selection	0.33	root		gt(test.t1.c2, 1)
 explain format = 'brief' select sum(t1.c1 in (select c1 from t2)) from t1;
 id	estRows	task	access object	operator info
 StreamAgg	1.00	root		funcs:sum(Column#13)->Column#11
-└─Projection	10000.00	root		cast(Column#10, decimal(65,0) BINARY)->Column#13
+└─Projection	10000.00	root		cast(Column#10, decimal(3,0) BINARY)->Column#13
   └─HashJoin	10000.00	root		CARTESIAN left outer semi join, other cond:eq(test.t1.c1, test.t2.c1)
     ├─IndexReader(Build)	10000.00	root		index:IndexFullScan
     │ └─IndexFullScan	10000.00	cop[tikv]	table:t2, index:c1(c1)	keep order:false, stats:pseudo
@@ -225,7 +225,7 @@ set @@session.tidb_opt_insubq_to_join_and_agg=0;
 explain format = 'brief' select sum(t1.c1 in (select c1 from t2)) from t1;
 id	estRows	task	access object	operator info
 StreamAgg	1.00	root		funcs:sum(Column#13)->Column#11
-└─Projection	10000.00	root		cast(Column#10, decimal(65,0) BINARY)->Column#13
+└─Projection	10000.00	root		cast(Column#10, decimal(3,0) BINARY)->Column#13
   └─HashJoin	10000.00	root		CARTESIAN left outer semi join, other cond:eq(test.t1.c1, test.t2.c1)
     ├─IndexReader(Build)	10000.00	root		index:IndexFullScan
     │ └─IndexFullScan	10000.00	cop[tikv]	table:t2, index:c1(c1)	keep order:false, stats:pseudo
@@ -241,7 +241,7 @@ HashJoin	10000.00	root		CARTESIAN left outer semi join, other cond:eq(1, test.t2
 explain format = 'brief' select sum(6 in (select c2 from t2)) from t1;
 id	estRows	task	access object	operator info
 StreamAgg	1.00	root		funcs:sum(Column#13)->Column#11
-└─Projection	10000.00	root		cast(Column#10, decimal(65,0) BINARY)->Column#13
+└─Projection	10000.00	root		cast(Column#10, decimal(3,0) BINARY)->Column#13
   └─HashJoin	10000.00	root		CARTESIAN left outer semi join, other cond:eq(6, test.t2.c2)
     ├─TableReader(Build)	10000.00	root		data:TableFullScan
     │ └─TableFullScan	10000.00	cop[tikv]	table:t2	keep order:false, stats:pseudo
@@ -809,7 +809,7 @@ Projection	1.00	root		Column#7
   └─HashAgg(Probe)	1.00	root		funcs:sum(Column#12)->Column#7
     └─HashJoin	1.00	root		CARTESIAN left outer join
       ├─HashAgg(Build)	1.00	root		group by:1, funcs:sum(Column#14)->Column#12
-      │ └─Projection	1.00	root		cast(Column#6, decimal(42,0) BINARY)->Column#14
+      │ └─Projection	1.00	root		cast(Column#6, decimal(20,0) BINARY)->Column#14
       │   └─MaxOneRow	1.00	root		
       │     └─Projection	1.00	root		Column#5
       │       └─TableDual	1.00	root		rows:1
@@ -831,7 +831,7 @@ Projection	1.00	root		Column#5
   │ └─TableReader	10000.00	root		data:TableFullScan
   │   └─TableFullScan	10000.00	cop[tikv]	table:t	keep order:false, stats:pseudo
   └─StreamAgg(Probe)	1.00	root		funcs:sum(Column#7)->Column#5
-    └─Projection	1.00	root		cast(Column#4, decimal(42,0) BINARY)->Column#7
+    └─Projection	1.00	root		cast(Column#4, decimal(20,0) BINARY)->Column#7
       └─TableDual	1.00	root		rows:1
 explain format = 'brief' select sum(a), (select sum(a)), count(a) from t group by b order by (select count(a));
 id	estRows	task	access object	operator info
@@ -853,7 +853,7 @@ admin reload opt_rule_blacklist;
 explain format = 'brief' select sum(t1.a) from t t1 join t t2 on t1.a=t2.a;
 id	estRows	task	access object	operator info
 StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
-└─Projection	12487.50	root		cast(test.t.a, decimal(25,0) BINARY)->Column#14
+└─Projection	12487.50	root		cast(test.t.a, decimal(3,0) BINARY)->Column#14
   └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.a, test.t.a)]
     ├─TableReader(Build)	9990.00	root		data:Selection
     │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.a))
@@ -864,7 +864,7 @@ StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
 explain format = 'brief' select sum(t1.b) from t t1 join t t2 on t1.b=t2.b;
 id	estRows	task	access object	operator info
 StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
-└─Projection	12487.50	root		cast(test.t.b, decimal(27,0) BINARY)->Column#14
+└─Projection	12487.50	root		cast(test.t.b, decimal(5,0) BINARY)->Column#14
   └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.b, test.t.b)]
     ├─TableReader(Build)	9990.00	root		data:Selection
     │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.b))
@@ -875,7 +875,7 @@ StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
 explain format = 'brief' select sum(t1.c) from t t1 join t t2 on t1.c=t2.c;
 id	estRows	task	access object	operator info
 StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
-└─Projection	12487.50	root		cast(test.t.c, decimal(30,0) BINARY)->Column#14
+└─Projection	12487.50	root		cast(test.t.c, decimal(8,0) BINARY)->Column#14
   └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.c, test.t.c)]
     ├─TableReader(Build)	9990.00	root		data:Selection
     │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.c))
@@ -886,7 +886,7 @@ StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
 explain format = 'brief' select sum(t1.d) from t t1 join t t2 on t1.d=t2.d;
 id	estRows	task	access object	operator info
 StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
-└─Projection	12487.50	root		cast(test.t.d, decimal(32,0) BINARY)->Column#14
+└─Projection	12487.50	root		cast(test.t.d, decimal(10,0) BINARY)->Column#14
   └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.d, test.t.d)]
     ├─TableReader(Build)	9990.00	root		data:Selection
     │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.d))
@@ -897,7 +897,7 @@ StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
 explain format = 'brief' select sum(t1.e) from t t1 join t t2 on t1.e=t2.e;
 id	estRows	task	access object	operator info
 StreamAgg	1.00	root		funcs:sum(Column#14)->Column#13
-└─Projection	12487.50	root		cast(test.t.e, decimal(41,0) BINARY)->Column#14
+└─Projection	12487.50	root		cast(test.t.e, decimal(20,0) BINARY)->Column#14
   └─HashJoin	12487.50	root		inner join, equal:[eq(test.t.e, test.t.e)]
     ├─TableReader(Build)	9990.00	root		data:Selection
     │ └─Selection	9990.00	cop[tikv]		not(isnull(test.t.e))