… cmd/explaintest/r/window_function.result
… cmd/explaintest/t/explain_easy.test
… executor/testdata/agg_suite_out.json
… expression/builtin.go
… parser/mysql/type.go
… planner/cascades/testdata/integration_suite_out.json
… planner/core/preprocess.go
… planner/core/testdata/integration_serial_suite_out.json
… planner/core/testdata/ordered_result_mode_suite_out.json
… planner/core/testdata/plan_suite_out.json
… planner/core/testdata/plan_suite_unexported_out.json
… planner/core/testdata/stats_suite_out.json