<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Problem Summary:

Currently the `sum` function will cast [MySQL integer types](https://dev.mysql.com/doc/refman/8.0/en/integer-types.html) `TINYINT, SMALLINT, MEDIUMINT, INT, BIGINT` to decimals with big precision, which is waste of CPU when pushed down to TiFlash.

### What is changed and how it works?

Change the argument decimal precision of `sum` to be just enough for the integer type.

It speeds up TiFlash in a query for TPCH-100 dataset:

```
select sum(L_LINENUMBER) from lineitem;
before: cast int to decimal(32, 0)   1.42s
after: cast int to decimal(10, 0)    1.1s
```

### Benchmark tests

We have conducted benchmark tests with this improvement when the `cast` is executed on TiDB/TiKV/TiFlash.

#### Results

Here's the summary of results. We expect that TiFlash is what really benefits from this improvement, because of it's implementation of different precision of `Decimal`. TiDB/TiKV's results are within acceptable experimental errors.

| engine  | sum(int) |        |        | sum(bigint) |        |        |
|---------|----------|--------|--------|-------------|--------|--------|
|         | before   | after  | %      | before      | after  | %      |
| TiDB    | 28.40s   | 28.79s | +1.4%  | 29.60s      | 28.26s | -4.5%  |
| TiFlash | 1.52s    | 1.18s  | -22.4% | 2.40s       | 1.63s  | -32.1% |
| TiKV    | 29.8s    | 29.09s | -2.4%  | 26.51s      | 26.75s | +0.9%  |

#### plan before PR

These are the plans before this PR respectively:

```sql
-- TiFlash
MySQL [tpch_100]> desc select sum(L_LINENUMBER) from lineitem;
+--------------------------------+--------------+-------------------+----------------+-----------------------------------------------------------------------+
| id                             | estRows      | task              | access object  | operator info                                                         |
+--------------------------------+--------------+-------------------+----------------+-----------------------------------------------------------------------+
| HashAgg_23                     | 1.00         | root              |                | funcs:sum(Column#22)->Column#18                                       |
| └─TableReader_25               | 1.00         | root              |                | data:ExchangeSender_24                                                |
|   └─ExchangeSender_24          | 1.00         | batchCop[tiflash] |                | ExchangeType: PassThrough                                             |
|     └─HashAgg_8                | 1.00         | batchCop[tiflash] |                | funcs:sum(Column#26)->Column#22                                       |
|       └─Projection_37          | 600037902.00 | batchCop[tiflash] |                | cast(tpch_100.lineitem.l_linenumber, decimal(32,0) BINARY)->Column#26 |
|         └─TableFullScan_22     | 600037902.00 | batchCop[tiflash] | table:lineitem | keep order:false                                                      |
+--------------------------------+--------------+-------------------+----------------+-----------------------------------------------------------------------+
6 rows in set (0.00 sec)

MySQL [tpch_100]> desc select sum(cast(L_LINENUMBER as signed)) from lineitem;
+--------------------------------+--------------+-------------------+----------------+------------------------------------------------------------------------------------------------+
| id                             | estRows      | task              | access object  | operator info                                                                                  |
+--------------------------------+--------------+-------------------+----------------+------------------------------------------------------------------------------------------------+
| HashAgg_23                     | 1.00         | root              |                | funcs:sum(Column#22)->Column#18                                                                |
| └─TableReader_25               | 1.00         | root              |                | data:ExchangeSender_24                                                                         |
|   └─ExchangeSender_24          | 1.00         | batchCop[tiflash] |                | ExchangeType: PassThrough                                                                      |
|     └─HashAgg_8                | 1.00         | batchCop[tiflash] |                | funcs:sum(Column#26)->Column#22                                                                |
|       └─Projection_37          | 600037902.00 | batchCop[tiflash] |                | cast(cast(tpch_100.lineitem.l_linenumber, bigint(22) BINARY), decimal(43,0) BINARY)->Column#26 |
|         └─TableFullScan_22     | 600037902.00 | batchCop[tiflash] | table:lineitem | keep order:false                                                                               |
+--------------------------------+--------------+-------------------+----------------+------------------------------------------------------------------------------------------------+
6 rows in set (0.00 sec)

-- TiKV

MySQL [tpch_100]> desc select sum(L_LINENUMBER) from lineitem;
+----------------------------+--------------+-----------+---------------------------------------------------------+------------------------------------------------------+
| id                         | estRows      | task      | access object                                           | operator info                                        |
+----------------------------+--------------+-----------+---------------------------------------------------------+------------------------------------------------------+
| StreamAgg_20               | 1.00         | root      |                                                         | funcs:sum(Column#23)->Column#18                      |
| └─IndexReader_21           | 1.00         | root      |                                                         | index:StreamAgg_8                                    |
|   └─StreamAgg_8            | 1.00         | cop[tikv] |                                                         | funcs:sum(tpch_100.lineitem.l_linenumber)->Column#23 |
|     └─IndexFullScan_19     | 600037902.00 | cop[tikv] | table:lineitem, index:PRIMARY(L_ORDERKEY, L_LINENUMBER) | keep order:false                                     |
+----------------------------+--------------+-----------+---------------------------------------------------------+------------------------------------------------------+

MySQL [tpch_100]> desc select sum(cast(L_LINENUMBER as signed)) from lineitem;
+----------------------------+--------------+-----------+---------------------------------------------------------+-------------------------------------------------------------------------------+
| id                         | estRows      | task      | access object                                           | operator info                                                                 |
+----------------------------+--------------+-----------+---------------------------------------------------------+-------------------------------------------------------------------------------+
| StreamAgg_20               | 1.00         | root      |                                                         | funcs:sum(Column#23)->Column#18                                               |
| └─IndexReader_21           | 1.00         | root      |                                                         | index:StreamAgg_8                                                             |
|   └─StreamAgg_8            | 1.00         | cop[tikv] |                                                         | funcs:sum(cast(tpch_100.lineitem.l_linenumber, bigint(22) BINARY))->Column#23 |
|     └─IndexFullScan_19     | 600037902.00 | cop[tikv] | table:lineitem, index:PRIMARY(L_ORDERKEY, L_LINENUMBER) | keep order:false                                                              |
+----------------------------+--------------+-----------+---------------------------------------------------------+-------------------------------------------------------------------------------+
4 rows in set (0.00 sec)

-- TiDB

MySQL [tpch_100]> set tidb_opt_cpu_factor=0; set tidb_opt_copcpu_factor=100000000000000000;
Query OK, 0 rows affected (0.00 sec)

Query OK, 0 rows affected (0.00 sec)

MySQL [tpch_100]> desc select sum(L_LINENUMBER) from lineitem;
+----------------------------+--------------+-----------+---------------------------------------------------------+-----------------------------------------------------------------------+
| id                         | estRows      | task      | access object                                           | operator info                                                         |
+----------------------------+--------------+-----------+---------------------------------------------------------+-----------------------------------------------------------------------+
| StreamAgg_9                | 1.00         | root      |                                                         | funcs:sum(Column#25)->Column#18                                       |
| └─Projection_26            | 600037902.00 | root      |                                                         | cast(tpch_100.lineitem.l_linenumber, decimal(32,0) BINARY)->Column#25 |
|   └─IndexReader_25         | 600037902.00 | root      |                                                         | index:IndexFullScan_24                                                |
|     └─IndexFullScan_24     | 600037902.00 | cop[tikv] | table:lineitem, index:PRIMARY(L_ORDERKEY, L_LINENUMBER) | keep order:false                                                      |
+----------------------------+--------------+-----------+---------------------------------------------------------+-----------------------------------------------------------------------+
4 rows in set (0.00 sec)

MySQL [tpch_100]> desc select sum(cast(L_LINENUMBER as signed)) from lineitem;
+----------------------------+--------------+-----------+---------------------------------------------------------+------------------------------------------------------------------------------------------------+
| id                         | estRows      | task      | access object                                           | operator info                                                                                  |
+----------------------------+--------------+-----------+---------------------------------------------------------+------------------------------------------------------------------------------------------------+
| StreamAgg_9                | 1.00         | root      |                                                         | funcs:sum(Column#25)->Column#18                                                                |
| └─Projection_26            | 600037902.00 | root      |                                                         | cast(cast(tpch_100.lineitem.l_linenumber, bigint(22) BINARY), decimal(43,0) BINARY)->Column#25 |
|   └─IndexReader_25         | 600037902.00 | root      |                                                         | index:IndexFullScan_24                                                                         |
|     └─IndexFullScan_24     | 600037902.00 | cop[tikv] | table:lineitem, index:PRIMARY(L_ORDERKEY, L_LINENUMBER) | keep order:false                                                                               |
+----------------------------+--------------+-----------+---------------------------------------------------------+------------------------------------------------------------------------------------------------+
4 rows in set (0.00 sec)
```

#### plans after PR

These are the plans after this PR:

```sql
-- TiFlash
MySQL [tpch_100]> desc select sum(L_LINENUMBER) from lineitem;
+--------------------------------+--------------+-------------------+----------------+-----------------------------------------------------------------------+
| id                             | estRows      | task              | access object  | operator info                                                         |
+--------------------------------+--------------+-------------------+----------------+-----------------------------------------------------------------------+
| HashAgg_23                     | 1.00         | root              |                | funcs:sum(Column#22)->Column#18                                       |
| └─TableReader_25               | 1.00         | root              |                | data:ExchangeSender_24                                                |
|   └─ExchangeSender_24          | 1.00         | batchCop[tiflash] |                | ExchangeType: PassThrough                                             |
|     └─HashAgg_8                | 1.00         | batchCop[tiflash] |                | funcs:sum(Column#26)->Column#22                                       |
|       └─Projection_37          | 600037902.00 | batchCop[tiflash] |                | cast(tpch_100.lineitem.l_linenumber, decimal(10,0) BINARY)->Column#26 |
|         └─TableFullScan_22     | 600037902.00 | batchCop[tiflash] | table:lineitem | keep order:false                                                      |
+--------------------------------+--------------+-------------------+----------------+-----------------------------------------------------------------------+
6 rows in set (0.01 sec)

MySQL [tpch_100]> desc select sum(cast(L_LINENUMBER as signed)) from lineitem;
+--------------------------------+--------------+-------------------+----------------+------------------------------------------------------------------------------------------------+
| id                             | estRows      | task              | access object  | operator info                                                                                  |
+--------------------------------+--------------+-------------------+----------------+------------------------------------------------------------------------------------------------+
| HashAgg_23                     | 1.00         | root              |                | funcs:sum(Column#22)->Column#18                                                                |
| └─TableReader_25               | 1.00         | root              |                | data:ExchangeSender_24                                                                         |
|   └─ExchangeSender_24          | 1.00         | batchCop[tiflash] |                | ExchangeType: PassThrough                                                                      |
|     └─HashAgg_8                | 1.00         | batchCop[tiflash] |                | funcs:sum(Column#26)->Column#22                                                                |
|       └─Projection_37          | 600037902.00 | batchCop[tiflash] |                | cast(cast(tpch_100.lineitem.l_linenumber, bigint(22) BINARY), decimal(20,0) BINARY)->Column#26 |
|         └─TableFullScan_22     | 600037902.00 | batchCop[tiflash] | table:lineitem | keep order:false                                                                               |
+--------------------------------+--------------+-------------------+----------------+------------------------------------------------------------------------------------------------+
6 rows in set (0.00 sec)

-- TiKV

MySQL [tpch_100]> desc select sum(L_LINENUMBER) from lineitem;
+----------------------------+--------------+-----------+---------------------------------------------------------+------------------------------------------------------+
| id                         | estRows      | task      | access object                                           | operator info                                        |
+----------------------------+--------------+-----------+---------------------------------------------------------+------------------------------------------------------+
| StreamAgg_20               | 1.00         | root      |                                                         | funcs:sum(Column#23)->Column#18                      |
| └─IndexReader_21           | 1.00         | root      |                                                         | index:StreamAgg_8                                    |
|   └─StreamAgg_8            | 1.00         | cop[tikv] |                                                         | funcs:sum(tpch_100.lineitem.l_linenumber)->Column#23 |
|     └─IndexFullScan_19     | 600037902.00 | cop[tikv] | table:lineitem, index:PRIMARY(L_ORDERKEY, L_LINENUMBER) | keep order:false                                     |
+----------------------------+--------------+-----------+---------------------------------------------------------+------------------------------------------------------+
4 rows in set (0.01 sec)

MySQL [tpch_100]> desc select sum(cast(L_LINENUMBER as signed)) from lineitem;
+----------------------------+--------------+-----------+---------------------------------------------------------+-------------------------------------------------------------------------------+
| id                         | estRows      | task      | access object                                           | operator info                                                                 |
+----------------------------+--------------+-----------+---------------------------------------------------------+-------------------------------------------------------------------------------+
| StreamAgg_20               | 1.00         | root      |                                                         | funcs:sum(Column#23)->Column#18                                               |
| └─IndexReader_21           | 1.00         | root      |                                                         | index:StreamAgg_8                                                             |
|   └─StreamAgg_8            | 1.00         | cop[tikv] |                                                         | funcs:sum(cast(tpch_100.lineitem.l_linenumber, bigint(22) BINARY))->Column#23 |
|     └─IndexFullScan_19     | 600037902.00 | cop[tikv] | table:lineitem, index:PRIMARY(L_ORDERKEY, L_LINENUMBER) | keep order:false                                                              |
+----------------------------+--------------+-----------+---------------------------------------------------------+-------------------------------------------------------------------------------+
4 rows in set (0.00 sec)

-- TiDB

MySQL [tpch_100]> desc select sum(L_LINENUMBER) from lineitem;
+----------------------------+--------------+-----------+---------------------------------------------------------+-----------------------------------------------------------------------+
| id                         | estRows      | task      | access object                                           | operator info                                                         |
+----------------------------+--------------+-----------+---------------------------------------------------------+-----------------------------------------------------------------------+
| StreamAgg_9                | 1.00         | root      |                                                         | funcs:sum(Column#25)->Column#18                                       |
…