@ngaut 2021-12-17
/run tpc-h