@fuzhe1989 2021-12-27
@siddontang TiDB will push the whole agg down to TiKV without appending any cast. This will also work for TiFlash in the future.

About unittest, this pr doesn't change any operator or function, the only thing it affects is the generated physical plan and I think explaintest is enough.